requirements.txt.txt
streamlit
google-generativeai
pillow
PyMuPDF